+++ config.py
@@ -1,4 +1,3 @@
-DATABASE = 'coriplus.sqlite'
 DEBUG = True
 SECRET_KEY = 'hin6bab8ge25*r=x&amp;+5$0kn=-#log$pt^#@vrqjld!^2ci@g*b'
 SITE_NAME = 'Cori+'