+++ migrate_0_6_to_0_7.py
@@ -1,0 +1,10 @@
+import sqlite3
+
+conn = sqlite3.connect('coriplus.sqlite')
+
+if __name__ == '__main__':
+    conn.executescript('''
+BEGIN TRANSACTION;
+  ALTER TABLE userprofile ADD COLUMN telegram TEXT;
+COMMIT;
+''')